+++ uspto/util/common.py
@@ -1,0 +1,24 @@
+# -*- coding: utf-8 -*-
+# (c) 2017 Andreas Motl <andreas@ip-tools.org>
+import sys
+import logging
+
+def to_list(obj):
+    """Convert an object to a list if it is not already one"""
+    # stolen from cornice.util
+    if not isinstance(obj, (list, tuple)):
+        obj = [obj, ]
+    return obj
+
+def boot_logging(options=None):
+    log_level = logging.INFO
+    if options and options.get('--debug'):
+        log_level = logging.DEBUG
+    setup_logging(level=log_level)
+
+def setup_logging(level=logging.INFO):
+    log_format = '%(asctime)-15s [%(name)-20s] %(levelname)-7s: %(message)s'
+    logging.basicConfig(
+        format=log_format,
+        stream=sys.stderr,
+        level=level)